update 21/11/2024 from levy2
