--- EventManagementSystem/Utils/ViewsPathServices/Implementations/BlogPathProvider.cs
@@ -1,16 +1,0 @@
-﻿using Web.Areas.Event.Controllers.Event;
-
-namespace Web.Utils.ViewsPathServices.Implementations
-{
-    public class BlogPathProvider : IPathProvider
-    {
-        public string GetViewsPath(object target)
-        {
-            return target switch
-            {
-                BlogController => "~/Views/Event",
-                _ => throw new ArgumentException($"Does not found {target.GetType().Name} in Account Area")
-            };
-        }
-    }
-}
