+++ EventManagementSystem/Utils/ViewsPathServices/Implementations/EventsPathProvider.cs
@@ -1,14 +1,15 @@
-﻿using Web.Areas.Event.Controllers.Event;
+﻿using Web.Views.Controllers.Event;
+using Web.Views.Controllers.Event;
 
 namespace Web.Utils.ViewsPathServices.Implementations
 {
-    public class DetailPathProvider : IPathProvider
+    public class EventsPathProvider : IPathProvider
     {
         public string GetViewsPath(object target)
         {
             return target switch
             {
-                DetailsController => "~~/Views/Event",
+                EventsController => "~/Views/Event",
                 _ => throw new ArgumentException($"Does not found {target.GetType().Name} in Account Area")
             };
         }
