--- EventManagementSystem/Utils/ViewsPathServices/Implementations/TermsPathProvider.cs
@@ -1,16 +1,0 @@
-﻿using Web.Areas.Event.Controllers.Event;
-
-namespace Web.Utils.ViewsPathServices.Implementations
-{
-    public class TermsPathProvider : IPathProvider
-    {
-        public string GetViewsPath(object target)
-        {
-            return target switch
-            {
-                TermsController => "~/Views/Event",
-                _ => throw new ArgumentException($"Does not found {target.GetType().Name} in Account Area")
-            };
-        }
-    }
-}
