+++ EventManagementSystem/Views/Event/Maps.cshtml
@@ -2,7 +2,7 @@
     <div class="row h-100">
         <div class="col text-white position-absolute" style="bottom: 10px">
             <h1 class="display-4">Event</h1>
-            <a asp-area="" asp-controller="Home" asp-action="Index" class="navbar-brand"><i class="bi bi-house-fill" style="font-size: 1.5rem;"></i></a><i style="font-size: 1.5rem">/ Blog</i>
+            <a asp-area="" asp-controller="Home" asp-action="Index" class="navbar-brand"><i class="bi bi-house-fill" style="font-size: 1.5rem;"></i></a><i style="font-size: 1.5rem">/ Map</i>
         </div>
     </div>
 </div>
@@ -49,7 +49,7 @@
     <div class="row g-4">
             <div class="col-md-6 col-lg-4">
                 <div class="card venue-card">
-                <a asp-area="Event" asp-controller="DetailMap" asp-action="DetailMap" class="DetailMap text-dark" style="text-decoration: none;">
+                <a asp-area="Event" asp-controller="Events" asp-action="DetailMap" class="DetailMap text-dark" style="text-decoration: none;">
 
                     <img src="https://vcdn1-dulich.vnecdn.net/2022/05/12/Hanoi2-1652338755-3632-1652338809.jpg?w=0&h=0&q=100&dpr=2&fit=crop&s=NxMN93PTvOTnHNryMx3xJw" class="card-img-top" alt="Australia Grand Hawk View">
                     <div class="card-body">
