+++ EventManagementSystem/Views/Shared/_Layout.cshtml
@@ -26,13 +26,13 @@
 							</a>
 						</li>
 						<li class="nav-item">
-							<a asp-area="Account" asp-controller="Register" asp-action="Register" class="nav-link">
+							<a asp-area="Account" asp-controller="Account" asp-action="Register" class="nav-link">
 								<i class="fas fa-user-plus"></i> <!-- Icon Font Awesome -->
 								Đăng kí
 							</a>
 						</li>
 						<li class="nav-item">
-							<a asp-area="Event" asp-controller="Map" asp-action="Maps" class="nav-link">
+							<a asp-area="Event" asp-controller="Events" asp-action="Maps" class="nav-link">
 								<i class="fas fa-map-marker-alt"></i> <!-- Icon Font Awesome -->
 								Địa điểm
 							</a>
@@ -49,7 +49,7 @@
 							</ul>
 						</li>
 						<li class="nav-item">
-							<a asp-area="Event" asp-controller="Event" asp-action="AllEvent" class="btn btn-primary text-white">Browse Events</a>
+							<a asp-area="Event" asp-controller="Events" asp-action="AllEvent" class="btn btn-primary text-white">Browse Events</a>
 						</li>
 					</ul>
 				</div>
@@ -70,10 +70,10 @@
 				<div class="col-md-3">
 					<h5>Thông tin</h5>
 					<ul class="list-unstyled">
-						<li><a a asp-area="Event" asp-controller="About" asp-action="About"  class="text-white">About</a></li>
+						<li><a asp-area="Event" asp-controller="Events" asp-action="About" class="text-white">About</a></li>
 						<li><a href="#" class="text-white">Events</a></li>
-						<li><a a asp-area="Event" asp-controller="Terms" asp-action="Terms" class="text-white">Terms & Conditions</a></li>
-						<li><a a asp-area="Event" asp-controller="Privacy" asp-action="Privacy" class="text-white">Privacy Policy</a></li>
+						<li><a asp-area="Event" asp-controller="Events" asp-action="Terms" class="text-white">Terms & Conditions</a></li>
+						<li><a asp-area="Event" asp-controller="Events" asp-action="Privacy" class="text-white">Privacy Policy</a></li>
 					</ul>
 				</div>
