+++ EventManagementSystem/Web.csproj
@@ -74,6 +74,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Areas\Account\Controllers\" />
     <Folder Include="Areas\Account\Views\Shared\" />
     <Folder Include="Areas\AdminEvents\Controllers\" />
     <Folder Include="Areas\AdminEvents\Data\" />
